Update select-player.js
Testing new code. Every time the `object` tag containing the `pitching-chart` svg loads (just switching between detail pages), it shows the first player in the table's pitches. This is because it still has the original `onload` event hard-coded in with the Liquid values. Now, it removes the listener and adds a new one to replace with correct pitches.