Replaced rating as class with separate attribute
Replaced using letter rating as a class with it being a separate attribute. It's used for defining the fill gradients. This change eliminates the need for using a `.replace()` call on the elements' `classList` to replace the existing rating with the new one. The CSS selector now checks for attribute content.