Updated incorrect pitching chart JS/Liquid logic
Updated incorrect function logic. It was actually not running as intended on initial load, as first player onload is a lefty, but setting `lefty` attribute to `'L'` instead of `true`. Worked in the `boolean` check in the JS, but was not the intended logic.

Now, the JS checks if `leftHanded` is a `string` or a `boolean` and acts accordingly.