Update update-pitch-chart.js
Re-worked logic once more. Because there are later uses of `leftHanded`, they aren't prepared for the possibility of it being a `string`. Now, there's a local variable called `leftHanded` that is the newly-renamed parameter `leftHandedInput` converted to a `boolean`. This is now the value that is checked throughout the code.