Update team-counts.scss
Needed make weird adjusts to padding of each row in the team counts box.

Due to the background color of the top row being the same color as the border, the border looks like it's part of the spacing. Therefore, needed to adjust. Also, needed to even out with bottom row and so entire box has the right height in the body grid.